requirements.txt
py-clob-client>=0.23.0
python-dotenv>=1.0.0
requests>=2.31.0
web3>=6.0.0
eth-account>=0.9.0
numpy>=1.24.0
annotated-types>=0.7.0
